pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-ticket-router"
version = "0.1.0"
description = "L1 Customer Support Ticket Router - an OpenEnv environment for email triage"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "ticket_router.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["ticket_router", "ticket_router.server"]
package-dir = { "ticket_router" = ".", "ticket_router.server" = "server" }
